// sanity.types.ts
export type ProductType = {
  _id: string;
  _type: "productType";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  name?: string;
  slug?: {
    _type: "slug";
    current: string;
  };
  images?: Array<{
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
    };
    hotspot?: {
      x: number;
      y: number;
      height: number;
      width: number;
    };
    crop?: {
      top: number;
      bottom: number;
      left: number;
      right: number;
    };
    alt?: string;
    _type: "image";
    _key: string;
  }>;
  description?: Array<
    | {
        children?: Array<{
          marks?: Array<string>;
          text?: string;
          _type: "span";
          _key: string;
        }>;
        style?: "normal" | "h1" | "h2" | "h3" | "h4" | "blockquote";
        listItem?: "bullet";
        markDefs?: Array<{
          href?: string;
          _type: "link";
          _key: string;
        }>;
        level?: number;
        _type: "block";
        _key: string;
      }
    | {
        asset?: {
          _ref: string;
          _type: "reference";
          _weak?: boolean;
        };
        hotspot?: {
          x: number;
          y: number;
          height: number;
          width: number;
        };
        crop?: {
          top: number;
          bottom: number;
          left: number;
          right: number;
        };
        alt?: string;
        _type: "image";
        _key: string;
      }
  >;
  price?: number;
  subcategories?: {
    _id: string;
    name: string;
  }[]; 
  scale?: string;
  stock?: number;
};